move table types, restrict dropzone to owners and admins, fix permission logic
